Quit writing MSW code in setup file
This doesn't work for some reason
