--- vitest.setup.ts
@@ -1,20 +1,0 @@
-import { afterAll, afterEach, beforeAll } from "vitest";
-import { setupServer } from "msw/node";
-import { http, HttpResponse } from "msw";
-
-export const restHandlers = [
-  http.get("/workers/getLocation", () => {
-    return HttpResponse.json({ ipData: { country: "CA" } });
-  }),
-];
-
-const server = setupServer(...restHandlers);
-
-// Start server before all tests
-beforeAll(() => server.listen({ onUnhandledRequest: "error" }));
-
-//  Close server after all tests
-afterAll(() => server.close());
-
-// Reset handlers after each test `important for test isolation`
-afterEach(() => server.resetHandlers());
